requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-QRcode==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.1
passlib==1.7.4
Pillow==9.0.1
pycparser==2.21
qrcode==7.3.1
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
WTForms==3.0.1
gunicorn==20.1.0
Babel==2.9.1
flask-paginate==2022.1.8
